+++ services/madoc-ts/src/frontend/site/pages/tasks/review-listing/review-listing-page.tsx
@@ -143,7 +143,7 @@ export function ReviewListingPage() {
     {
       keepPreviousData: true,
       getFetchMore: lastPage => {
-        if (lastPage.pagination.totalPages === lastPage.pagination.page) {
+        if (lastPage.pagination.totalPages === 0 || lastPage.pagination.totalPages === lastPage.pagination.page) {
           return undefined;
         }
         return {